+++ ark-common/src/main/java/au/org/theark/core/service/ArkSettingServiceImpl.java
@@ -58,7 +58,7 @@ public class ArkSettingServiceImpl implements IArkSettingService {
     }
 
     @Override
-    public List<Setting> searchPageableSettings(Setting setting, int first, int count) {
+    public List<Setting> searchPageableSettings(Setting setting, long first, long count) {
         return iArkSettingDao.searchPageableSettings(setting, first, count);
     }
