+++ ark-lims/src/main/java/au/org/theark/lims/model/dao/IBioTransactionDao.java
@@ -56,7 +56,7 @@ public interface IBioTransactionDao {
 	 * @param count
 	 * @return List<au.org.theark.core.model.lims.entity.BioTransaction> given the criteria
 	 */
-	public List<BioTransaction> searchPageableBioTransactions(BioTransaction bioTransaction, int first, int count);
+	public List<BioTransaction> searchPageableBioTransactions(BioTransaction bioTransaction, long first, long count);
 
 	/**
 	 * Create a LIMS BioTransaction based on the supplied bioTransaction
