+++ ark-lims/src/main/java/au/org/theark/lims/service/ILimsSubjectService.java
@@ -42,5 +42,5 @@ public interface ILimsSubjectService {
 	 * @param count
 	 * @return
 	 */
-	public abstract List<LinkSubjectStudy> searchPageableSubjects(LimsVO limsVoCriteria, List<Study> studyList, int first, int count);
+	public abstract List<LinkSubjectStudy> searchPageableSubjects(LimsVO limsVoCriteria, List<Study> studyList, long first, long count);
 }
